% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrypop.R
\name{countrypop}
\alias{countrypop}
\title{Countrypop function}
\usage{
countrypop(x)
}
\arguments{
\item{a}{"Country"}
}
\description{
Returns a graph of a country's population over time.
}
\examples{
countrypop("United States of America")
}
