% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{FIFA World Cup Hosts and Attendance 1930-2022}
\format{
A data frame with 22 observations
\describe{
\item{Year}{Year world cup was played}
\item{Hosts}{Host country}
\item{Matches}{Total matches played that year}
\item{Totalattendance}{Total attendance at all matches}
\item{Averageattendance}{Average attendance across all matches that year}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
FIFA World Cup Hosts and Attendance 1930-2022
}
\keyword{datasets}
