% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Population of United Nations countries for 1950-2020}
\format{
A data frame with 235 observations
}
\usage{
WorldPopulation
}
\description{
Populations in thousands
}
\keyword{datasets}
